requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-cleanup==3.1.0
django-cool-pagination==0.3.0
django-cors-headers==2.5.0
django-heroku==0.3.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
whitenoise==4.1.2
